+++ src/Equinor.ProCoSys.Completion.WebApi/appsettings.json
@@ -74,5 +74,19 @@
     "Scoped": {}
   },
   "UseAzureAppConfiguration": true,
-  "AdapterPartitions": "TSTG"
+  "StartTieImport": false,
+  "TieImport": {
+    "AdapterTieUri": "",
+    "AzureClientId": "",
+    "AzureCertificateName": "",
+    "AzureKeyVaultUrl": "",
+    "AzureTenantId": "",
+    "AzureTieApiId": "",
+    "AdapterParallelMessageHandling": true,
+    "AdapterMessageChunkSize": 1,
+    "AdapterIdleTimeBetweenBatch": 1000,
+    "AdapterIdleTimeOnNoMessages": 30000,
+    "AdapterSites": "",
+    "AdapterApplication": ""
+  }
 }
